#693: Added dictionary tests in JSON
